requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
python-telegram-bot==20.3
requests==2.31.0
aiohttp==3.8.5
tensorflow==2.10.1
google-api-python-client==2.122.0
google-auth==2.29.0